Move more daemon setup logic from lib.rs to main.rs
