+++ tests/test_add.rs
@@ -1,1 +1,4 @@
+use pueue_daemon_lib::run;
+
+#[test]
 fn test_add() {}
